[REFACTOR] excpetion도 도메인별로 분리
