+++ main.py
@@ -1,7 +1,6 @@
 from fastapi import FastAPI
 from api.available_room import router as available_router
-from exception import custom_exception_handler
-from exception.common_exception import BaseCustomException
+from exception import custom_exception_handler, BaseCustomException
 
 app = FastAPI()
 app.include_router(available_router)
