+++ tests/exception/test_date_validator.py
@@ -1,6 +1,6 @@
 import pytest
 from utils.validate.common.date_validator import validate_date_format, validate_date_not_past
-from exception.common_exception import InvalidDateFormatError, PastDateNotAllowedError
+from exception.common.date_exception import InvalidDateFormatError, PastDateNotAllowedError
 from datetime import datetime, timedelta
 
 def test_validate_date_valid():
