+++ tests/exception/test_response_validator.py
@@ -1,6 +1,6 @@
 import pytest
 from utils.validate.common.response_validator import validate_response_rooms
-from exception.common_exception import ResponseMismatchError
+from exception.common.response_exception import ResponseMismatchError
 from models.dto import RoomKey
 from pathlib import Path
 import json
